chore(e2e): e2e support file renaming
