+++ packages/suite-desktop-core/e2e/support/pageActions/recoveryActions.ts
